+++ shared/src/main/scala/com/nawforce/pkgforce/modifiers/Modifier.scala
@@ -14,7 +14,6 @@
 package com.nawforce.pkgforce.modifiers
 
 import com.nawforce.pkgforce.diagnostics.Duplicates.IterableOps
-import com.nawforce.pkgforce.diagnostics.{LogEntryContext, ModifierLogger}
 import com.nawforce.runtime.parsers.CodeParser
 import com.nawforce.runtime.parsers.CodeParser.ParserRuleContext
 import io.github.apexdevtools.apexparser.ApexParser.{